nutriplan_experiments/requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Training & Optimization
peft>=0.4.0
bitsandbytes>=0.39.0
deepspeed>=0.9.0

# Evaluation Metrics
nltk>=3.8
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=3.1.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.4.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Knowledge Graph
networkx>=3.1
rdflib>=6.3.0

# Web & API (optional)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
